+++ tests/testthat/test_occu.R
@@ -118,7 +118,7 @@ test_that("occu can fit models with covariates",{
 
   fitted <- fitted(fm)
   expect_equivalent(fitted, structure(c(0.5738, 0.5014, 0.4318, 0.38581, 0.50171, 0.53764,
-0.46563, 0.40283, 0.39986, 0.79928), .Dim = c(5L, 2L)), tol = 1e-5)
+0.46563, 0.40283, 0.39986, 0.79928), dim = c(5L, 2L)), tol = 1e-5)
 
   # methods
   gp <- getP(fm)
@@ -217,10 +217,10 @@ test_that("occu handles offsets",{
   umf <- unmarkedFrameOccu(y = y, siteCovs = siteCovs, obsCovs = obsCovs)
   fm <- occu(~ o1 + o2 ~ offset(x), data = umf)
   expect_equivalent(coef(fm),
-                     structure(c(9.74361, 0.44327, -0.14683, 0.44085), .Names = c("psi(Int)",
+                     structure(c(9.74361, 0.44327, -0.14683, 0.44085), names = c("psi(Int)",
 "p(Int)", "p(o1)", "p(o2)")), tol = 1e-5)
   fm <- occu(~ o1 + offset(o2) ~ offset(x), data = umf)
-  expect_equivalent(coef(fm), structure(c(8.59459, 0.97574, -0.3096), .Names = c("psi(Int)",
+  expect_equivalent(coef(fm), structure(c(8.59459, 0.97574, -0.3096), names = c("psi(Int)",
 "p(Int)", "p(o1)")), tol=1e-5)
 
 })