+++ tests/testthat/test_occuPEN.R
@@ -100,7 +100,7 @@ test_that("occuPEN can fit models with covariates",{
 
   fitted <- fitted(fm)
   expect_equivalent(fitted, structure(c(0.5738, 0.5014, 0.4318, 0.38581, 0.50171, 0.53764,
-0.46563, 0.40283, 0.39986, 0.79928), .Dim = c(5L, 2L)), tol = 1e-5)
+0.46563, 0.40283, 0.39986, 0.79928), dim = c(5L, 2L)), tol = 1e-5)
 
   expect_error(occuPEN_CV(~ o1 + o2 ~ x, data = umf, k=15))
   fmCV <- occuPEN_CV(~ o1 + o2 ~ x, data = umf)
@@ -171,10 +171,31 @@ test_that("occuPEN can handle offsets",{
   umf <- unmarkedFrameOccu(y = y, siteCovs = siteCovs, obsCovs = obsCovs)
   fm <- occuPEN(~ o1 + o2 ~ offset(x), data = umf)
   expect_equivalent(coef(fm),
-                     structure(c(9.74361, 0.44327, -0.14683, 0.44085), .Names = c("psi(Int)",
+                     structure(c(9.74361, 0.44327, -0.14683, 0.44085), names = c("psi(Int)",
 "p(Int)", "p(o1)", "p(o2)")), tol = 1e-5)
   fm <- occuPEN(~ o1 + offset(o2) ~ offset(x), data = umf)
-  expect_equivalent(coef(fm), structure(c(8.59459, 0.97574, -0.3096), .Names = c("psi(Int)",
+  expect_equivalent(coef(fm), structure(c(8.59459, 0.97574, -0.3096), names = c("psi(Int)",
 "p(Int)", "p(o1)")), tol=1e-5)
 
+})
+
+test_that("occuPEN MPLE works with missing values in y", {
+  # issue #54 thanks @RaphBnrd
+  y <- data.frame(matrix(c(
+    0, 1, 0, 0, 1, 1,
+    0, 0, 0, 0, 0, 0,
+    1, 1, 1, 1, 0, 0,
+    1, 1, 0, 1, 0, NA
+  ), nrow=4, byrow=TRUE))
+
+  siteCovs <- data.frame(x = c(-1, 1, -1, -1))
+
+  umf <- unmarkedFrameOccu(
+    y = y, siteCovs = siteCovs, obsCovs = NULL
+  )
+
+  formula <- ~1 ~ x
+
+  mod <- expect_warning(occuPEN(formula, data=umf, pen.type="MPLE"))
+  expect_is(mod, "unmarkedFitOccuPEN")
 })